feature: IP选择器支持以hostId作为主机唯一标识进行交互 #978
代码优化
