--- src/backend/job-manage/service-job-manage/src/main/java/com/tencent/bk/job/manage/common/util/StringParamUtil.java
@@ -1,54 +1,0 @@
-/*
- * Tencent is pleased to support the open source community by making BK-JOB蓝鲸智云作业平台 available.
- *
- * Copyright (C) 2021 THL A29 Limited, a Tencent company.  All rights reserved.
- *
- * BK-JOB蓝鲸智云作业平台 is licensed under the MIT License.
- *
- * License for BK-JOB蓝鲸智云作业平台:
- * --------------------------------------------------------------------
- * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and associated
- * documentation files (the "Software"), to deal in the Software without restriction, including without limitation
- * the rights to use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the Software, and
- * to permit persons to whom the Software is furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all copies or substantial portions of
- * the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO
- * THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF
- * CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS
- * IN THE SOFTWARE.
- */
-
-package com.tencent.bk.job.manage.common.util;
-
-import org.apache.commons.lang3.StringUtils;
-
-import java.util.Arrays;
-import java.util.Collections;
-import java.util.List;
-
-/**
- * 模块内共享的字符型参数处理工具类
- */
-public class StringParamUtil {
-
-    /**
-     * 使用常见的分号、逗号、换行、空格等字符分割字符串
-     *
-     * @param str 目标字符串
-     * @return 分割后的字符串列表
-     */
-    public static List<String> splitByNormalSeparator(String str) {
-        if (str == null) {
-            return Collections.emptyList();
-        }
-        if (StringUtils.isBlank(str)) {
-            return Collections.singletonList(str);
-        }
-        return Arrays.asList(str.split("[;,；，\\n\\s|\\u00A0]+"));
-    }
-
-}
